% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_mod_muni_select.R
\name{mod_mod_muni_select_ui}
\alias{mod_mod_muni_select_ui}
\alias{mod_mod_muni_select_server}
\title{mod_mod_muni_select_ui and mod_mod_muni_select_server}
\usage{
mod_mod_muni_select_ui(id)

mod_mod_muni_select_server(input, output, session)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
